refactor(ATT-424): split attractap API god class into modular files
Break the 1429-line api.cpp god file into cohesive translation units
and extract a real OtaUpdater collaborator, with zero change to the
public API:: surface consumed by the display screens and application.

- api.cpp (1429 -> 298): core spine — lifecycle, message dispatch,
  outbound messaging, callback setters
- api_auth.cpp: registration, authentication, device-name handshake
- api_resources.cpp: resource list parsing + session/door/flow/billing commands
- api_cards.cpp: card auth data exchange + new-card enrollment
- api_projects.cpp: paginated projects-of-user request/response
- api_forms.cpp: usage-form request parsing + submission serialization
- ota/ota_updater.{hpp,cpp}: OtaUpdater class owns the OTA state machine;
  esp_ota_ops.h/esp_partition.h pulled out of the public api.hpp header

All four board envs build green; behavior preserved (pure move + collaborator
extraction with injected sendMessage/logger/callbacks).
